src/libs/base_libs.py
import os
import io
import sys
import cv2
import math
import time
import yaml
import glob
import copy
import json
import torch
# import onnx
import shutil
#import uvicorn
import logging 
import datetime
import traceback
import torchvision
import numpy as np
# import pandas as pd
from math import ceil
from pathlib import Path
from loguru import logger
import onnxruntime as ort
from PIL import Image as Im
# from functools import reduce
from typing import Optional, List, Union
from itertools import product as product
from pydantic import BaseModel, StrictBool
from datetime import date as dtdate
from datetime import timedelta, timezone
